Got rid of more globals Cleaned up some extra unised variables Converted most settings from int to long Added code to support processor speed and tradeoff

+++ dTV/dTV/DI_BlendedClip.c
@@ -92,7 +92,7 @@
 // still see too many Weave artifacts ("venetion blinds").  It's best to try everything
 // else first.
 
-	int		BlcMinimumClip = -15;				// currently -100 .. 100
+	long		BlcMinimumClip = -15;				// currently -100 .. 100
 
 // "Pixel Motion Sensitivity" slider:  This determines how sensitive we are to motion.
 // Motion is calculated as the maximum absolute change in luma from the previous field
@@ -101,7 +101,7 @@
 // of venetiaon blinds that can occur with sudden scene changes.  This value is
 // calculated separately for each pixel.
 
-	UINT	BlcPixelMotionSense = 17;
+	long	BlcPixelMotionSense = 17;
 
 // "Recent Motion Sensitivity" slider:  This increases the tendency to use Clip based
 // upon an n-period Exponential Moving Average of the recent motion.  Recent motion
@@ -110,7 +110,7 @@
 // does not attempt to do 3:2 pulldown I believe the motion values could be of assistance
 // in the routines that do.  
 
-	int		BlcRecentMotionSense = 0;		// current -100 .. 100)		
+	long		BlcRecentMotionSense = 0;		// current -100 .. 100)		
 
 // "Motion Average Period" slider:  This sets the period of the moving average for Recent
 // Motion Sensitivity.  
@@ -123,7 +123,7 @@
 
 //			X_new_avg = ( X_old_avg * (n-1) + 2 * X) / (n+1)
 
-	UINT	BlcMotionAvgPeriod = 20;		// currently 1..200
+	long	BlcMotionAvgPeriod = 20;		// currently 1..200
 
 // "Pixel Comb Sensitivity" slider:  This determines how sensitive we are to the current
 // comb factor of each pixel.  I used a simplified comb factor C = abs(2*W - H - L)/2,
@@ -132,17 +132,17 @@
 // Motion Sense seem to be the two main things to play with to get good results.  Generally,
 // increase one of these if you get Weave artifacts and decrease one if you get BOB artifacts.
   
-	UINT	BlcPixelCombSense = 27;
+	long	BlcPixelCombSense = 27;
 
 // "Recent Comb Senseitivity" slider:  Operates like the Recent Motion slider but operates
 // on the average Comb Factor.
 
-	UINT	BlcRecentCombSense = 0;
+	long	BlcRecentCombSense = 0;
 
 // "Comb Average Period" slider: Sets the period of the Comb exponential moving average.
 // See the comments on "Motion Average Period".
 
-	UINT	BlcCombAvgPeriod = 20;			// currently 1.200
+	long	BlcCombAvgPeriod = 20;			// currently 1.200
 
 // "Skip High Comb Frames" slider:  I added this one in the hopes that it could help to
 // skip a frame in the event of a sudden flash attack on a rapid scene change or maybe
@@ -150,7 +150,7 @@
 // a chance to experiment with it yet.  It will give very ugly results if you set it 
 // too high.
 
-	UINT	BlcHighCombSkip = 10;			// larger values skip more
+	long	BlcHighCombSkip = 10;			// larger values skip more
 
 // "Skip Low Motion Frames" slider:  This also is just experimental an probably of low
 // value.  The idea here is that any frame with sufficiently low change from the previous
@@ -158,12 +158,12 @@
 // normal use.  NOTE - This slider (but not parm) will soon be replaced by the
 // Vertical Smoothing slider.
 
-	UINT	BlcLowMotionSkip = 0;			// larger values skip more
+	long	BlcLowMotionSkip = 0;			// larger values skip more
 
 // "Vertical Smoothing" slider: Sets a smoothing constant to smooth between the even
 // and odd lines.  Not yet implemented, but the INI parm is there.
 
-	UINT    BlcVerticalSmoothing = 0;
+	long    BlcVerticalSmoothing = 0;
 
 // "Use Interpolated BOB instead of Clip" check box.  For those who don't like the
 // Clipped Weave, this will change it to an Interpolated Bob.  All other blending and
@@ -189,10 +189,10 @@
 
 // Other global values, not user parms:
 
-	UINT	BlcAverageMotions[5][2] = {0};  // reserved
-	UINT	BlcTotalAverageMotion = 0;
-	UINT	BlcAverageCombs[5][2] = {0};	// reserved
-	UINT	BlcTotalAverageComb = 0;
+	long	BlcAverageMotions[5][2] = {0};  // reserved
+	long	BlcTotalAverageMotion = 0;
+	long	BlcAverageCombs[5][2] = {0};	// reserved
+	long	BlcTotalAverageComb = 0;
 	BOOL	BlcWantsToFlip;
